+++ src/winpriv.h
@@ -76,6 +76,7 @@ extern bool show_charinfo;
 extern void toggle_charinfo(void);
 extern void toggle_vt220(void);
 extern char * fontpropinfo(void);
+extern void show_config_log(void);
 
 extern bool title_settable;
 extern bool support_wsl;